refactor:more semantic component
